Fix RenderUtil primitives, registry lookup via getOrThrow, DamageSources, bare player refs, ConfigScreen click API

+++ src/main/java/dev/zprestige/prestige/client/managers/DamageManager.java
@@ -61,7 +61,7 @@ public class DamageManager implements MC {
         double damage = ((impact * impact + impact) / 2 * 7 * (6 * 2) + 1);
 
         damage = getDamageForDifficulty(damage);
-        damage = DamageUtil.getDamageLeft(player, damage, getMc().world.getDamageSources().generic(), (float) player.getArmor(), (float) player.getAttributeInstance(EntityAttributes.ARMOR_TOUGHNESS).getValue());
+        damage = DamageUtil.getDamageLeft(player, (float) damage, getMc().world.getDamageSources().generic(), (float) player.getArmor(), (float) player.getAttributeInstance(EntityAttributes.ARMOR_TOUGHNESS).getValue());
         damage = resistanceReduction(player, damage);
 
         damage = blastProtReduction(player, damage);
@@ -195,7 +195,7 @@ public class DamageManager implements MC {
         damage = resistanceReduction(player, damage);
 
         // Reduce by armour
-        damage = DamageUtil.getDamageLeft(player, damage, getMc().world.getDamageSources().generic(), (float) player.getArmor(), (float) player.getAttributeInstance(EntityAttributes.ARMOR_TOUGHNESS).getValue());
+        damage = DamageUtil.getDamageLeft(player, (float) damage, getMc().world.getDamageSources().generic(), (float) player.getArmor(), (float) player.getAttributeInstance(EntityAttributes.ARMOR_TOUGHNESS).getValue());
 
         // Reduce by enchants
         damage = blastProtReduction(player, damage);
